# Tweak the logic about generating 'RasterCacheKey' of the type 'kLayerChildren'