@flar 2022-04-15
The referenced issue https://github.com/flutter/flutter/issues/101872 doesn't seem to have any test cases to demonstrate the problem it assumes. It does point to problems in https://github.com/flutter/flutter/issues/101597 though which are due to a more fundamental problem in how Flutter manages animations. They need to fix that and so many other cases will improve as well.

This solution, while it might help a little with some test cases (which? it doesn't cite any), only targets a particular (and alleged) symptom of the problem whereas solution is available to the framework and app developers that will obviate the need for this PR and others like it.

Let's fix these issues from the top down.

See examples presented in these comments:

- https://github.com/flutter/flutter/issues/101597#issuecomment-1099936511
- https://github.com/flutter/flutter/issues/101597#issuecomment-1099942361