@ColdPaleLight 2022-04-15
@flar 
I think about the two test cases of the issue https://github.com/flutter/flutter/issues/101597, @jonahwilliams  hopes that developers can solve it without manually adding `RepaintBoundary`.

case 1 

|  solution   | result  |
|  ----  | ----  |
| without change  | not work |
| manually adding `RepaintBoundary` | works |
| with engine PR https://github.com/flutter/engine/pull/32676    | not work,  `DisplayList`'s `unique_id` keeps changing |
| with flutter PR https://github.com/flutter/flutter/pull/101952   | not work,  `DisplayListLayer`'s `unique_id` keeps changing |
| with engine PR https://github.com/flutter/flutter/pull/101952 and flutter PR https://github.com/flutter/flutter/pull/101952   | not work,  `DisplayList`'s `unique_id` keeps changing |




case 2

|  solution   | result  |
|  ----  | ----  |
| without change  | not work |
| manually adding `RepaintBoundary` | works |
| with engine PR https://github.com/flutter/engine/pull/32676    | not work,  `DisplayList`'s `unique_id` keeps changing |
| with flutter PR https://github.com/flutter/flutter/pull/101952   | not work,  `DisplayListLayer`'s `unique_id` keeps changing |
| with engine PR https://github.com/flutter/flutter/pull/101952 and flutter PR https://github.com/flutter/flutter/pull/101952   | works |