@flar 2022-04-15
The point I'm making is that there are other reasons for them to add a RepaintBoundary and we should work on dealing with that (perhaps having the standard AnimatedFoo classes all do it for them or something). Making this change in the engine for that reason means we are providing fewer reasons for them to make the right change.

Also, this change addresses one stability problem. If the children are not a picture or dl layer, but they are stable, we don't get any benefit from this change. The animating widgets need to ensure the stability of their children whether or not it is a simple picture layer and since they don't do that, this PR has only solved the problem for animators with picture children.

The RepaintBoundary isn't there to provide us a caching opportunity. That's not why it was created. It was created to reduce work in the framework so it doesn't have to visit as many nodes in the tree to rebuild on every frame. That saves a bunch of work in the "Build" phase of the tree that we don't even see.

At some point we added the picture cache which took advantage of the isolation and stability that the RepaintBoundary provides. We'll still get that benefit, and we'll get it more effectively if the framework were to work on its sub-tree stability better.