@flar 2022-04-15
I do like the idea that caching a DisplayListLayer should be equivalent to caching its embedded DisplayList directly.

Is there a simpler solution with, say, a layer override. Some way that we say "layer cache yourself" and it says "you mean, cache my DL"? We wouldn't have to mess with rearranging the cache IDs, just implement "when you try to cache a DL layer, it will catch that and cache its DL instead".

The only tricky part is if a parent layer says "cache my children" and the child is singular, but do we already catch that?