@jonahwilliams 2022-04-15
@ColdPaleLight https://github.com/flutter/flutter/pull/101952 has been updated to keep the child picture stable for the AnimatedSwitcher