@jonahwilliams 2022-04-15
> The point I'm making is that there are other reasons for them to add a RepaintBoundary and we should work on dealing with that (perhaps having the standard AnimatedFoo classes all do it for them or something). Making this change in the engine for that reason means we are providing fewer reasons for them to make the right change.

I don't disagree with you here, but I think that as long as we are going to attempt to cache children we should do it correctly. Or bail out and avoid populating a cache we'll never hit