@ColdPaleLight 2022-04-18
@jonahwilliams @flar 
I filed another PR here https://github.com/flutter/engine/pull/32738 to implement Option 1 mentioned in the Issue. Perhaps that is more suitable for solving linked issue.