@jonahwilliams 2022-04-18
This change seems fairly straightforward compared to a breaking API change in dart:ui.