@ColdPaleLight 2022-04-21
> 
Hi @flar 
Sorry, but I'm afraid that you might have some misunderstanding of what this PR is trying to do.

This PR wants to solve the problem that `kChildrenLayer` will fail If DisplayListLayer is unstable but DisplayLit is stable. See the linked issue for details https://github.com/flutter/flutter/issues/101872 .

I thought maybe  "DLLayer caching == DL caching" would be nice, but that mechanism doesn't seem to do anything for the issue we would like to solve. Please let me know if i missed something, thank you very much.