@flar 2022-04-21
There are a number of outstanding issues right now to change our raster cache and they are starting to blend together in my head. The top concern for me is that if we have to do this much work on the raster caching code, and if each solution involves new mechanisms on top of old mechanisms, then maybe we are too deep in the grass and need to stand up and re-assess.

In particular, in your [table](https://github.com/flutter/engine/pull/32676#issuecomment-1100007712) above, https://github.com/flutter/flutter/pull/101952 is deemed as not solving the problem, but I think it is the right solution here. It seems to be an ongoing work and it might just solve the problem on its own if we wait for it to settle.

I strongly support that any animation:

- isolate itself from disturbing the tree above it. A RepaintBoundary above it can do that, but the "AnimatingFoo" paradigm side steps that by have a builder phase in its RO that rebuilds without disturbing the tree. That's a win.
- not rebuild its children if they don't need to rebuild. https://github.com/flutter/flutter/pull/101952 seems to be trying to address that, but I'm not sure it is quite there yet (I need to look at it more closely, but I'm not a framework expert).

We have tools to address those principles above and we should make use of them as much as possible before complicating the raster cache much further.

Having said that, this PR is not a huge complication, but it does add a few more steps and a couple more classes. We should wait and see what we can accomplish with https://github.com/flutter/flutter/pull/101952 as this PR alone doesn't solve the problem itself per the above table. If we can't get there with 101952 alone, then this PR seems like a good way to help with that.