@ColdPaleLight 2022-04-27
> Overall this looks pretty clean. I've made a couple of suggestions that might improve performance. Do we have any benchmarks that measure the cost of caching (prepare/draw/etc) with these more complicated keys?

I tested it with the following code:

1. With this PR
``` c++
  auto start = fml::TimePoint::Now();
  RasterCacheKey::Map<int> map;
  auto hash_function = map.hash_function();
  SkMatrix matrix = SkMatrix::I();
  for (int i = 0; i < 100000; i++) {
    RasterCacheKeyID first = RasterCacheKeyID(foo, RasterCacheKeyType::kLayer);
    RasterCacheKeyID second = RasterCacheKeyID(bar, RasterCacheKeyType::kLayer);
    RasterCacheKeyID third =
      RasterCacheKeyID({first, second}, RasterCacheKeyType::kLayerChildren);
    RasterCacheKey key(third, matrix);
    hash_function(key);
  }
  auto end = fml::TimePoint::Now();
  FML_DLOG(INFO) << "cost " << (end - start).ToNanoseconds();
```
2. Without this PR：
``` c++
  auto start = fml::TimePoint::Now();
  RasterCacheKey::Map<int> map;
  auto hash_function = map.hash_function();
  SkMatrix matrix = SkMatrix::I();
  for (int i = 0; i < 100000; i++) {
    RasterCacheKey key(RasterCacheKeyID({foo, bar}), RasterCacheKeyType::kLayerChildren, matrix);
    hash_function(key);
  }
  auto end = fml::TimePoint::Now();
  FML_DLOG(INFO) << "cost " << (end - start).ToNanoseconds();
```

With this PR：
total 315889084 ns， Average time per key is 315 ns.

Without this PR:
total 136134995 ns, Average time per key is 136 ns.

So I think this PR will not bring significant performance issues.